Correctly handle empty datatypes
DataTypes can be defined which do nothing but extend an xs simple type. In this situation an XSD is cleaner to not define these datatypes.
The code was designed to do this but it wasnt implementing it correctly.

Add test to prove defining a mandatoryString datatype which provides no restrictions just renders the XSD with xs:string